% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as.numeric.factor}
\alias{as.numeric.factor}
\title{Convert the levels of a factor to number
A convenience function for converting factor levels into numeric}
\usage{
\method{as.numeric}{factor}(factor)
}
\arguments{
\item{factor}{A \code{factor} vector to convert to numeric.}
}
\value{
A \code{numeric} vector of the factor levels.
}
\description{
Convert the levels of a factor to number
A convenience function for converting factor levels into numeric
}
