% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{subsetSharedCohortsAndSamples}
\alias{subsetSharedCohortsAndSamples}
\title{Subset meta estimate data to shared cohorts and shared per cohort
  samples}
\usage{
subsetSharedCohortsAndSamples(metaestimateData)
}
\arguments{
\item{metaestimateData}{A \code{list} containing a per classifier
survival predictions as well as ground truth survival data in
the item names 'survival'.}
}
\value{
The metaestimate \code{list} with only cohorts common to all
   classifiers and only samples common to all classifier cohrots.
}
\description{
Subset meta estimate data to shared cohorts and shared per cohort
  samples
}
